% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(n, x, y)
}
\arguments{
\item{n}{number of bootstrap replicates}

\item{x}{variable}

\item{y}{variable}
}
\value{
Correlation between x and y \cr
The bootstrap estimate of correlation \cr
The bootstrap estimation of standard error\cr
Bias \cr
Confidence interval
}
\description{
This function uses the boostrap method to estimate the correlation, standard error, bias and confidence interval. The function takes a .txt, which contains the x and y variables, as an input argument.
}
\author{
Katrine Eriksen and Katrine Bach \cr
Department of mathematics and computer science (IMADA) \cr
University of Southern Denmark \cr
\email{kater13@student.sdu.dk} and \email{kabac13@student.sdu.dk}
}
\keyword{bootstrap,}
\keyword{correlation}
\keyword{error,}
\keyword{standard}

