% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffon.R
\name{buffon}
\alias{buffon}
\title{Buffon's Neddle experiment}
\usage{
buffon(N,l,d)
}
\arguments{
\item{N}{number of needles to be thrown on the floor.}

\item{l}{length of the needle.}

\item{d}{distance between the stripes on the floor.}
}
\value{
P(hit)\cr
pi
}
\description{
This function estimates the probability of a needle to hit a line, P(hit), and pi by performing Buffon's experiment.
}
\examples{
buffon(1000,1,1)
}
\author{
Katrine Eriksen and Katrine Bach \cr
Department of mathematics and computer science (IMADA) \cr
University of Southern Denmark \cr
\email{kater13@student.sdu.dk} and \email{kabac13@student.sdu.dk}
}
\keyword{buffon}

