% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.probability.R
\name{chi.probability}
\alias{chi.probability}
\title{Chi square probability}
\usage{
chi.probability(x, df, n)
}
\arguments{
\item{x}{the point where we want to calculate the probability.}

\item{df}{degree of freedom}

\item{n}{number of random numbers for each variable.}
}
\description{
This function applies the Monte Carlo method to generate a chi-squared distribution.
}
\examples{
chi.probability(10,4,100)
}
\author{
Katrine Eriksen and Katrine Bach \cr
Department of mathematics and computer science (IMADA) \cr
University of Southern Denmark \cr
\email{kater13@student.sdu.dk} and \email{kabac13@student.sdu.dk}
}
\keyword{bootstrap}

