% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densi.R
\name{densi}
\alias{densi}
\title{Estimation of Density}
\usage{
densi(x,d=NULL,h=NULL, method="naive")
}
\arguments{
\item{x}{numeric value}

\item{d}{a point that is evaluatet to obtain its density}

\item{h}{bin width}

\item{method}{the method for density estimation. It can be "naive" or "kernel".}
}
\description{
This function estimates density using the naive estimator or gaussian kernel.
}
\examples{
densi(faithful$eruptions)
}
\author{
Katrine Eriksen and Katrine Bach \cr
Department of mathematics and computer science (IMADA) \cr
University of Southern Denmark \cr
\email{kater13@student.sdu.dk} and \email{kabac13@student.sdu.dk}
}
\keyword{density}

