% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoodnessOfFit.R
\name{GoodnessOfFit}
\alias{GoodnessOfFit}
\title{Goodness of Fit}
\usage{
GoodnessOfFit(x,p)
}
\arguments{
\item{x}{observed values}

\item{p}{the expected probability for all outcomes}
}
\description{
This function perfomes chi-squared test and goodness-of-fit test.
}
\author{
Katrine Eriksen and Katrine Bach \cr
Department of mathematics and computer science (IMADA) \cr
University of Southern Denmark \cr
\email{kater13@student.sdu.dk} and \email{kabac13@student.sdu.dk}
}
\keyword{GOF.}
\keyword{goodness-of-fit}

