% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChain.R
\name{MarkovChain}
\alias{MarkovChain}
\title{Markov Chain}
\usage{
MarkovChain(p,k,n)
}
\arguments{
\item{p}{Transition probability matrix}

\item{k}{integer indicating the initial state of the chain}

\item{n}{number of simulated steps}
}
\value{
Returns a table of the stationary probabilities of each state.
}
\description{
This function simulates the stationary distribution of a Markov Chain.
}
\author{
Katrine Eriksen and Katrine Bach \cr
Department of mathematics and computer science (IMADA) \cr
University of Southern Denmark \cr
\email{kater13@student.sdu.dk} and \email{kabac13@student.sdu.dk}
}
\keyword{Markov}

