% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixRegression.R
\name{MatrixRegression}
\alias{MatrixRegression}
\title{Matrix Regression}
\usage{
MatrixRegression(formula)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}
}
\value{
Returns a linear regression model
}
\description{
This function uses matrix regression to fit a simple or multiple linear regression model.
}
\author{
Katrine Eriksen and Katrine Bach \cr Department of mathematics and
  computer science (IMADA) \cr University of Southern Denmark \cr
  \email{kater13@student.sdu.dk} and \email{kabac13@student.sdu.dk}
}
\keyword{matrix}
\keyword{regression}

